Improve support for non-ASCII inputs using unicode
Replace uses of the win32 ToAscii() API with ToUnicode(),
and pass a 16-bit unicode code point in the BARRIER_MSG_KEY
wParam instead of a codepage-specific 8-bit value.

This is still broken for any unicode beyond the Basic
Multilingual Plane (BMP) because it assumes that each
16-bit wchar_t value is a single character, whereas Windows
will produce surrogate pairs for higher planes.

To get support beyond the BMP, the output of ToUnicode()
would need to be decoded from UTF-16 to UTF-32 and the
full 21-bit value be encoded into the wParam. The dead-key
logic will need to operate on full UTF-32 units instead
of UTF-16.